tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-auto-tracks-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masonry: masonry-auto-tracks getComputedStyle()</title>
<link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<style>
  #target {
    font-size: 40px;
  }
</style>
<script>
// <track-breadth>
// <track-breadth> = <length-percentage> | <flex> | min-content | max-content | auto
test_computed_value("masonry-auto-tracks", "1px");
test_computed_value("masonry-auto-tracks", "calc(10px + 0.5em)", "30px");
test_computed_value("masonry-auto-tracks", "calc(10px - 0.5em)", "0px");
test_computed_value("masonry-auto-tracks", "4%");
test_computed_value("masonry-auto-tracks", "5fr");
test_computed_value("masonry-auto-tracks", "min-content");
test_computed_value("masonry-auto-tracks", "max-content");
test_computed_value("masonry-auto-tracks", "auto");

// minmax( <inflexible-breadth> , <track-breadth> )
// <inflexible-breadth> = <length-percentage> | min-content | max-content | auto
test_computed_value("masonry-auto-tracks", "minmax(1px, 5fr)");
test_computed_value("masonry-auto-tracks", "minmax(calc(10px + 0.5em), max-content)", "minmax(30px, max-content)");
test_computed_value("masonry-auto-tracks", "minmax(calc(10px - 0.5em), max-content)", "minmax(0px, max-content)");
test_computed_value("masonry-auto-tracks", "minmax(4%, auto)");
test_computed_value("masonry-auto-tracks", "minmax(min-content, calc(10px + 0.5em))", "minmax(min-content, 30px)");
test_computed_value("masonry-auto-tracks", "minmax(auto, 4%)");

// fit-content( <length-percentage> )
test_computed_value("masonry-auto-tracks", "fit-content(1px)");
test_computed_value("masonry-auto-tracks", "fit-content(calc(10px + 0.5em))", "fit-content(30px)");
test_computed_value("masonry-auto-tracks", "fit-content(calc(10px - 0.5em))", "fit-content(0px)");
test_computed_value("masonry-auto-tracks", "fit-content(4%)");

// 0
test_computed_value("masonry-auto-tracks", "0px");
test_computed_value("masonry-auto-tracks", "0%");
test_computed_value("masonry-auto-tracks", "0fr");
test_computed_value("masonry-auto-tracks", "minmax(auto, 0%)");
test_computed_value("masonry-auto-tracks", "fit-content(0px)");

// <track-size>+
test_computed_value("masonry-auto-tracks", "1px 2px 3px 0px");
test_computed_value("masonry-auto-tracks", "fit-content(1px) minmax(2px, 3px) 4px");
</script>
</body>
</html>
